+++ src/lib/desktopCard.svelte
@@ -1,233 +1,139 @@
 <script>
-    export let data;
-    import {breakpoint, platform} from '../mediaquery.svelte'; // importing mediaqueries!
-    import ButtonComponent from '$lib//button.svelte';
-    import Brief from '$lib/briefForm.svelte';
+    export let persons;
+    import ButtonComponent from "$lib//button.svelte";
+    import pfp from "$lib/index.js";
+
+    function checkAvatarImage(avatar) {
+        if (avatar.includes("https")) {
+            return true;
+        }
+
+        return false;
+    }
 </script>
 
-{#if $breakpoint == "xs"}
+
 <style>
-    .container{
+    .cards-container {
         display: flex;
-        justify-content: center;
-        align-items: center;
-        height: 100vh;
-        background-color: rgb(31, 31, 31);
-        color: rgb(145, 213, 240);
-    }
+        flex-wrap: wrap;
+        gap: 16px;
 
-    .businessCard{
-        position: relative;
-        width: 90%;
-        height: 300px;
-        background-color: #252525;
-        border-radius: 8px;
-        border: 2px solid #91d5f025;
-        display: flex;
-        justify-content: center;
-        margin: auto;
-        box-shadow: 4px 8px 12px 8px rgba(0,0,0,0.35);
+        & .container {
+            flex: 1 1 calc(15%);
+            box-sizing: border-box;
+        }
 
-        & .leftAligned{
-            position: absolute;
-            left: 0;
-            height: 100%;
-            width: 50%;
-            
-            & img{
-                height: 100%;
-                width: 100%;
-                border-radius: 8px 0 0 8px; 
-                border-right: 1px solid #42C9FF;
+        & .container:nth-child(even) {
+            & .businessCard {
+                transform: rotate(-0.5deg);
             }
         }
 
-        & .rightAligned{
-            margin-left: 1em;
+        & .container:nth-child(odd) {
+            & .businessCard {
+                transform: rotate(0.5deg);
+            }
+        }
 
-            & h1{
+        & .businessCard {
+            position: relative;
+            width: 525px;
+            height: 350px;
+            background-color: #fdf3e2;
+            border-radius: 8px;
+            border: 2px solid #3b3b3b;
+            display: flex;
+            justify-content: center;
+            margin: auto;
+            box-shadow: 4px 8px 12px 8px rgba(0, 0, 0, 0.25);
+
+            & .leftAligned {
                 position: absolute;
-                top: 1em;
-                margin: .2em;
-                color: #42C9FF;
-                font-size: 1em;
+                left: 0;
+                height: 100%;
+                width: 40%;
+
+                & img {
+                    height: 100%;
+                    width: 100%;
+                    border-radius: 6px 0 0 6px;
+                    border-right: 2px solid #3b3b3b;
+                    object-fit: cover;
+                }
             }
 
-            & h2{
-                position: absolute;
-                top: 3em;
-                margin: .2em;
-                font-size: 1em;
-                color: #91D5F0;
-            }
+            & .rightAligned {
+                margin-left: -5em;
 
-            & p {
-                position: absolute;
-                top: 5.5em;
-                margin: .2em;
-                color: #FFFFFF;
+                & h1 {
+                    position: absolute;
+                    top: 0;
+                    margin: 0;
+                    color: #002b3b;
+                    font-size: 1.45em;
+                    font-weight: 700;
+                    text-transform: uppercase;
+                }
+
+                & h2 {
+                    position: absolute;
+                    top: 1.5em;
+                    margin: 0;
+                    color: #080052;
+                    font-size: 1.25em;
+                    font-weight: 600;
+                }
+
+                & p {
+                    position: absolute;
+                    top: 4em;
+                    margin: 0;
+                    color: #080052;
+                    width: 25ch;
+                }
+
+                & a {
+                    position: absolute;
+                    bottom: 0;
+                    right: 0;
+                    background-color: #140a3d;
+                    color: white;
+                    font-size: 16px;
+                    padding: 16px;
+                    border: none;
+                    cursor: pointer;
+                    text-align: center;
+                    width: 108px;
+                    border-radius: 24px 0 6px;
+                    box-shadow: 0px 0px 8px 2px rgba(0, 0, 0, 0.2);
+                    height: 47px;
+                    grid-column: 1;
+                    text-decoration: none;
+                }
             }
         }
     }
 </style>
-{/if}  
 
-{#if $breakpoint == "s"}
-<style>
-    .container{
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        height: 100vh;
-        background-color: rgb(31, 31, 31);
-        color: rgb(145, 213, 240);
-    }
-
-    .businessCard{
-        position: relative;
-        width: 90%;
-        height: 300px;
-        background-color: #252525;
-        border-radius: 8px;
-        border: 2px solid #91d5f025;
-        display: flex;
-        justify-content: center;
-        margin: auto;
-        box-shadow: 4px 8px 12px 8px rgba(0,0,0,0.35);
-
-        & .leftAligned{
-            position: absolute;
-            left: 0;
-            height: 100%;
-            width: 45%;
-            
-            & img{
-                height: 100%;
-                width: 100%;
-                border-radius: 8px 0 0 8px; 
-                border-right: 1px solid #42C9FF;
-            }
-        }
-
-        }
-        
-        & .rightAligned{
-            margin-left: 1em;
-
-            & h1{
-                position: absolute;
-                top: 1em;
-                margin: .2em;
-                color: #42C9FF;
-            }
-
-            & h2{
-                position: absolute;
-                top: 3em;
-                margin: .2em;
-                font-size: 1.2em;
-                color: #91D5F0;
-            }
-
-            & p {
-                position: absolute;
-                top: 5.5em;
-                margin: .2em;
-                color: #FFFFFF;
-            }
-        }
-    }
-</style>
-{/if}  
-
-{#if $breakpoint == "m"}
-<style>
-    .container{
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        height: 100vh;
-        background-color: rgb(31, 31, 31);
-        color: rgb(145, 213, 240);
-    }
-
-    .businessCard{
-        position: relative;
-        width: 700px;
-        height: 400px;
-        background-color: #252525;
-        border-radius: 8px;
-        border: 2px solid #91d5f025;
-        display: flex;
-        justify-content: center;
-        margin: auto;
-        box-shadow: 4px 8px 12px 8px rgba(0,0,0,0.35);
-
-        & .leftAligned{
-            position: absolute;
-            left: 0;
-            height: 100%;
-            width: 45%;
-            
-            & img{
-                height: 100%;
-                width: 100%;
-                border-radius: 8px 0 0 8px; 
-                border-right: 1px solid #42C9FF;
-            }
-        }
-
-        & .figure{
-            position: absolute;
-            bottom: 0;
-            z-index: 1;
-        }
-
-        & .rightAligned{
-            margin-left: 10em;
-
-            & h1{
-                position: absolute;
-                top: 1.2em;
-                margin: .25em;
-                color: #42C9FF;
-            }
-
-            & h2{
-                position: absolute;
-                top: 3.25em;
-                margin: .25em;
-                font-size: 1.25em;
-                color: #91D5F0;
-            }
-
-            & p {
-                position: absolute;
-                top: 6.25em;
-                margin: .25em;
-                color: #FFFFFF;
-            }
-        }
-    }
-</style>
-{/if}    
-
-{#each data.squadD as person}
-<div class="container">
-    <article class="businessCard">
-        <div class="leftAligned">
-            {#if person.avatar}
-                <img alt="Foto van {person.name}" src="{person.avatar}">
-            {:else}
-                <img alt="standard icon"
-                     src="../../static/images/profile-icon.png"/>
-            {/if}
+<section class="cards-container">
+    {#each persons as person}
+        <div class="container">
+            <article class="businessCard">
+                <div class="leftAligned">
+                    {#if person.avatar}
+                        <img src={person.avatar} alt=" " />
+                    {:else}
+                        <img src={pfp} alt=" " />
+                    {/if}
+                </div>
+                <div class="rightAligned">
+                    <h1>{person.name} {person.prefix} {person.surname}</h1>
+                    <h2>klas : {person.squad_id}</h2>
+                    <p>{person.bio}</p>
+                    <ButtonComponent />
+                </div>
+            </article>
         </div>
-        <div class="rightAligned">
-            <h1> <a href="/{person.id}"> {person.name} </a> </h1>
-            <p> klas : {person.squad_id} </p>
-            <ButtonComponent> </ButtonComponent>
-        </div>
-    </article>
-</div>
-{/each}
+    {/each}
+</section>
